Dialogue System
Built a simple dialogue system by using basic UI components - a button and a text labels.
The names for gaming components and various classes and variables used in the programs  are a bit ambiguous and can be changed appropriately.
The dialog system uses a button for 2 purposes - to start the dialog and also as a 'continue' or a 'next' button.
This system also uses Scriptable Objects to store data - the name of the  narrator and the lines.
The end of the dialogue has not been added yet, as it can be implemented very easily.