Implemented dialogue system in Infirmary Exit
A dialogue system has been implemented in the Infirmary Exit scene where there is an NPC with whom the MC interacts automatically and only once.

A basic capsule is taken as the NPC.

The dialogue gets triggered with the help of an edge collider set onto the capsule.

A few minor changes such as adding a collider for the ground, adding a composite collider to the ground, and adding a sorting layer for the NPCs has been done, because it was necessary to implement the player movement and the triggering mechanism.

The scripts written for the dialogue system have been revised and rewritten, to add the functionality of switching between scriptable objects.